Switched DataContainer Dimensions to 'size_t' which will vary based on how DREAM3D is compiled. The writing to HDF5 files is still with 64 bit integers. This part of the code compiles with a few hundred warnings. Added also some sanity checks for some of the data file readers to make sure we are not going to try and read more data that can fit into our address space.

+++ Source/EbsdLib/TSL/H5AngImporter.cpp
@@ -123,7 +123,7 @@ H5AngImporter::~H5AngImporter()
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-void H5AngImporter::getDims(int &x, int &y)
+void H5AngImporter::getDims(int64_t &x, int64_t &y)
 {
   x = xDim;
   y = yDim;
@@ -143,7 +143,7 @@ void H5AngImporter::getResolution(float &x, float &y)
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-int H5AngImporter::importFile(hid_t fileId, int z, const std::string &angFile)
+int H5AngImporter::importFile(hid_t fileId, int64_t z, const std::string &angFile)
 {
   herr_t err = -1;
   setCancel(false);
