+++ Source/EbsdLib/TSL/H5AngImporter.h
@@ -13,8 +13,8 @@
  * list of conditions and the following disclaimer in the documentation and/or
  * other materials provided with the distribution.
  *
- * Neither the name of Michael A. Groeber, Michael A. Jackson, the US Air Force, 
- * BlueQuartz Software nor the names of its contributors may be used to endorse 
+ * Neither the name of Michael A. Groeber, Michael A. Jackson, the US Air Force,
+ * BlueQuartz Software nor the names of its contributors may be used to endorse
  * or promote products derived from this software without specific prior written
  * permission.
  *
@@ -79,7 +79,7 @@ class EbsdLib_EXPORT H5AngImporter : public EbsdImporter
      * @param index The slice index for the file
      * @param angFile The absolute path to the input .ang file
      */
-    int importFile(hid_t fileId, int index, const std::string &angFile);
+    int importFile(hid_t fileId, int64_t index, const std::string &angFile);
 
     /**
      * @brief Writes the phase data into the HDF5 file
@@ -102,7 +102,7 @@ class EbsdLib_EXPORT H5AngImporter : public EbsdImporter
      * @param x Number of X Voxels (out)
      * @param y Number of Y Voxels (out)
      */
-    virtual void getDims(int &x, int &y);
+    virtual void getDims(int64_t &x, int64_t &y);
 
     /**
      * @brief Returns the x and y resolution of the voxels
@@ -116,8 +116,8 @@ class EbsdLib_EXPORT H5AngImporter : public EbsdImporter
 
 
   private:
-    int xDim;
-    int yDim;
+    int64_t xDim;
+    int64_t yDim;
     float xRes;
     float yRes;
